+++ CHANGELOG.md
@@ -9,6 +9,7 @@ You can also check [on GitHub](https://github.com/nextcloud/news/releases), the 
 ### Changed
 - Add feature to Group starred Items per Feed
 - Add `occ` commands for importing/exporting articles
+- Add importing/exporting articles to frontend settings
 
 ### Fixed
 - Special characters may be displayed incorrectly when full text is enabled